libc/wchar: Remove the unnecessary cast

+++ libs/libc/wchar/lib_wctob.c
@@ -57,6 +57,6 @@ int wctob(wint_t wc)
       return EOF;
     }
 
-  return wctomb(pmb, wc) == 1 ? (int)pmb[0] : EOF;
+  return wctomb(pmb, wc) == 1 ? pmb[0] : EOF;
 }
 #endif
